Implementation/accountservice/requirements.txt
sqlalchemy
Flask
bcrypt
psycopg2
uwsgi
validate_email
pika
pyjwt
requests